requirements.txt
selenium
webdriver-manager
pydantic
python-docx